+++ src/ParcelManager.DTO/Parcels/ParcelAddDto.cs
@@ -6,7 +6,7 @@
         public string ParcelNumber { get; set; } = default!;
         public string RecipientName { get; set; } = default!;
         public string DestinationCountry { get; set; } = default!;
-        public float Weight { get; set; } = default!;
-        public float Price { get; set; } = default!;
+        public decimal Weight { get; set; } = default!;
+        public decimal Price { get; set; } = default!;
     }
 }
